fix!: Move backwards-compatibility hacks out of tooltip.js
Move accessor properties for Blockly.Tooltip.visible and .DIV
from core/tooltip.js to core/main.js.
